Split off EnumIterator
Reduces dependency on tools.h
