+++ src/util/tools.h
@@ -163,43 +163,6 @@ std::string fallbackString(const std::string& first, const std::string& fallback
 /// \return return the (unsigned int) hash value
 unsigned int stringHash(std::string_view str);
 
-/// \brief Iterator over values of a sequential enum between begin and end
-template <typename C, C beginVal, C endVal>
-class EnumIterator {
-private:
-    using val_t = std::underlying_type_t<C>;
-    val_t val;
-
-public:
-    explicit EnumIterator(const C& f)
-        : val(static_cast<val_t>(f))
-    {
-    }
-    EnumIterator()
-        : val(static_cast<val_t>(beginVal))
-    {
-    }
-
-    EnumIterator operator++()
-    {
-        ++val;
-        return *this;
-    }
-
-    C operator*() { return static_cast<C>(val); }
-
-    EnumIterator begin() const { return *this; }
-    EnumIterator end() const
-    {
-        static const EnumIterator endIter = EnumIterator(endVal);
-        return endIter;
-    }
-    bool operator!=(const EnumIterator& i) const
-    {
-        return val != i.val;
-    }
-};
-
 /// \brief Get value of map, iff not key is not in map return defval
 template <typename K, typename V>
 V getValueOrDefault(const std::vector<std::pair<K, V>>& m, const K& key, const V& defval)
